Avoid deepcopying all register contents.
Machine state needs to be persisted during choice point creation, and later restorations must not
change the stored contents. The use of deepcopy is ideal in this case but also wasteful, because
all the contents of registers are also deep-copied. Cells referenced by the stored state may be mutated
because they'll be undone when the trail is unwound.

This solution simply creates a new object with almost all the same elements, even recursively cloning
environments (which may be unnecessary). I've tried simply using 'from copy import copy', which doesn't
work and I don't want to dig into why not (:

This change improves performance ~7x, from 2.09s to run test_grammar.py to 0.28s!