Make put_term iterative instead of recursive.
This change introduces some inefficiences, because we try to reuse structs already seen, either in the current
or previous args. The conflict avoidance algorithm is not ready to deal with structs, so one there's a conflict
the struct is moved multiple times. Also, since we're now treating structs as reusable stuff, just like vars,
some spurious MOVs are written, even after the last time the struct is used, just like happens already with vars.

This change now allows grammar.py to run! It still doesn't work, but we're now closer to it with no stack overflow.